+++ crates/rbuilder-config/Cargo.toml
@@ -10,6 +10,7 @@ exclude.workspace = true
 
 [dependencies]
 serde.workspace = true
+serde_with.workspace = true
 toml.workspace = true
 eyre.workspace = true
 tracing-subscriber = { version = "0.3.18", features = ["env-filter", "json"] }
