+++ crates/test-relay/src/main.rs
@@ -4,8 +4,8 @@ use metrics::spawn_metrics_server;
 use rbuilder::{
     beacon_api_client::Client,
     mev_boost::{MevBoostRelaySlotInfoProvider, RelayClient},
-    utils::tracing::{setup_tracing_subscriber, LoggerConfig},
 };
+use rbuilder_config::LoggerConfig;
 use relay::spawn_relay_server;
 use std::net::SocketAddr;
 use tokio_util::sync::CancellationToken;
@@ -81,12 +81,12 @@ async fn main() -> eyre::Result<()> {
 
     let global_cancellation = CancellationToken::new();
 
-    let config = LoggerConfig {
+    let logger_config = LoggerConfig {
         env_filter: cli.rust_log,
         log_json: cli.log_json,
         log_color: false,
     };
-    setup_tracing_subscriber(config)?;
+    logger_config.init_tracing()?;
 
     spawn_metrics_server(cli.metrics_address);
